moving utils and interfaces
