+++ packages/discussions/src/index.ts
@@ -5,3 +5,8 @@ export * from "./posts";
 export * from "./channels";
 export * from "./reactions";
 export * from "./types";
+
+export * from "./utils/posts";
+export * from "./utils/channels";
+export * from "./utils/reactions";
+export * from "./utils/platform";
